Custom adapter still giving trouble; still not sure why Added button in the template
