+++ app/src/main/java/king/bas/a1_comp3275/ActionBarAndMenu.java
@@ -5,6 +5,8 @@ import android.support.design.widget.FloatingActionButton;
 import android.support.design.widget.Snackbar;
 import android.support.v7.app.AppCompatActivity;
 import android.support.v7.widget.Toolbar;
+import android.view.Menu;
+import android.view.MenuInflater;
 import android.view.View;
 
 public class ActionBarAndMenu extends AppCompatActivity {
@@ -17,6 +19,14 @@ public class ActionBarAndMenu extends AppCompatActivity {
         setSupportActionBar(toolbar);
 
         getSupportActionBar().setDisplayHomeAsUpEnabled(true);
+
+    }
+
+    @Override
+    public boolean onCreateOptionsMenu(Menu menu) {
+        MenuInflater inflater = getMenuInflater();
+        inflater.inflate(R.menu.action, menu);
+        return super.onCreateOptionsMenu(menu);
     }
 
 }
