+++ app/src/main/res/layout/item.xml
@@ -13,7 +13,7 @@
 <!--for descriptions-->
     <LinearLayout
         android:orientation="vertical"
-        android:layout_width="fill_parent"
+        android:layout_width="189dp"
         android:layout_marginLeft="5dp"
         android:layout_height="wrap_content">
 
@@ -31,5 +31,11 @@
             android:layout_marginTop="5dp"/>
 
     </LinearLayout>
+
+    <Button
+        android:layout_width="wrap_content"
+        android:layout_height="63dp"
+        android:text="Details"
+        android:id="@+id/btn_deets" />
 
 </LinearLayout>
