+++ app/src/main/res/values/strings.xml
@@ -58,11 +58,20 @@
         <item>LCD Television</item>
     </string-array>
 
+<!--array for prices-->
     <string-array name="itemS_prices">
         <item>$5,459</item>
         <item>$6,00</item>
         <item>$5,459..00</item>
         <item>$7,501.00</item>
+    </string-array>
+
+    <!--// array for description-->
+    <string-array name="items_description">
+        <item>A Brand new Dell Laptop. A Bit outdated but still functional</item>
+        <item>A used iPhone 6. Functional and no scratches</item>
+        <item>A Brand new iPad</item>
+        <item>A Samsung Flat Smart TV</item>
     </string-array>
 
 </resources>
